Refactoring code

+++ src/Controller/ThemeController.php
@@ -11,36 +11,66 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Serializer\Encoder\JsonDecode;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 
 #[Route('/api/theme', name: 'app_theme_')]
 class ThemeController extends AbstractController
 {
     private $themeRepository;
-    private $tokenStorage;
     private $entityManager;
+    private $requestStack;
+    private $validator;
+    private $userRepository;
+    private $accountRepository;
     public function __construct(
         ThemeRepository $themeRepository,
-        TokenStorageInterface $tokenStorage,
         EntityManagerInterface $entityManager,
+        RequestStack $requestStack,
+        ValidatorInterface $validator,
+        UserRepository $userRepository,
+        AccountRepository $accountRepository,
     ) {
         $this->themeRepository = $themeRepository;
-        $this->tokenStorage = $tokenStorage;
         $this->entityManager = $entityManager;
+        $this->requestStack = $requestStack;
+        $this->validator = $validator;
+        $this->userRepository = $userRepository;
+        $this->accountRepository = $accountRepository;
     }
     #[Route('/admin/ajout', name: 'admin_add', methods: ['POST'])]
-    public function addTheme(
-        Request $request,
-        ValidatorInterface $validator,
-    ): JsonResponse {
-        $currentUser = $this->tokenStorage->getToken()->getUser()->getRoles()[0];
-        if ($currentUser != 'ROLE_ADMIN') {
+    public function addTheme(Request $request): JsonResponse
+    {
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        //récupération du role de l'utilisateur
+        $currentUserRole = $session->get('user_role')[0];
+
+        // si l'utilisateur n'a pas le role admin renvoi une erreur
+        if ($currentUserRole != 'ROLE_ADMIN') {
             return new JsonResponse(["message" => "Vous n'êtes pas autorisé à ajouter des thèmes"], JsonResponse::HTTP_FORBIDDEN);
         }
         // récupération des données de création d'un thème
         $data = json_decode($request->getContent(), true);
+
+        // vérifier si le theme existe déjà
+        $existingTheme = $this->themeRepository->findOneBy(['name' => $data['name']]);
+        if ($existingTheme) {
+            //le theme existe déjà, renvoi une réponse d'erreur
+            return new JsonResponse(['message' => 'Le theme existe déjà'], JsonResponse::HTTP_CONFLICT);
+        }
 
         // création d'un nouveau thème
         $theme = new Theme();
@@ -48,7 +78,7 @@ class ThemeController extends AbstractController
         $theme->setValue($data['value']);
 
         //valider les données du theme
-        $errors = $validator->validate($theme);
+        $errors = $this->validator->validate($theme);
         if (count($errors) > 0) {
             // il y a des erreurs de validation
             $errorsMessages = [];
@@ -56,13 +86,6 @@ class ThemeController extends AbstractController
                 $errorMessage = $error->getMessage();
             }
             return new JsonResponse(['message' => $errorMessage], JsonResponse::HTTP_BAD_REQUEST);
-        }
-
-        // vérifier si le theme existe déjà
-        $existingTheme = $this->themeRepository->findOneBy(['name' => $data['name']]);
-        if ($existingTheme) {
-            //le theme existe déjà, renvoi une réponse d'erreur
-            return new JsonResponse(['message' => 'Le theme existe déjà'], JsonResponse::HTTP_CONFLICT);
         }
 
         //enregistrer le theme dans la base de données
@@ -76,23 +99,35 @@ class ThemeController extends AbstractController
     public function getTheme(
         $id,
         Request $request,
-        AccountRepository $accountRepository,
     ): JsonResponse {
+
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
+        }
+
         // récupération de l'identifiant du compte passé en paramètre
         $accountId = $request->query->get('accountId');
+
         // recherche du compte dans la base de données
-        $account = $accountRepository->find($accountId);
+        $account = $this->accountRepository->find($accountId);
 
         // contrôle si le compte existe
-        if(!$account){
+        if(!$account) {
             return new JsonResponse(['message' => 'Le compte n\'existe pas'], JsonResponse::HTTP_NOT_FOUND);
         }
 
-        // récupération de l'identifiant de l'utilisateur courant par le token
-        $currentUser = $this->tokenStorage->getToken()->getUser()->getId();
-
         // on vérifie que l'identifiant utilisateur du compte correspond bien à l'identifiant de l'utilisateur courant dans le token
-        if($account->getUser()->getId() != $currentUser) {
+        if($account->getUser()->getId() != $user->getId()) {
             return new JsonResponse(['message' => 'Vous n\'avez pas accès à ce compte'], JsonResponse::HTTP_FORBIDDEN);
         }
 
@@ -146,26 +181,39 @@ class ThemeController extends AbstractController
 
         return new JsonResponse($questionsWithAnswers, JsonResponse::HTTP_OK);
     }
+
     #[Route('/game/getAll', name: 'gameGetAll', methods: ['GET'])]
     public function getAllAccountID(
-        AccountRepository $accountRepository,
-        Request $request): JsonResponse
-    {
+        Request $request
+    ): JsonResponse {
+
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
+        }
 
         // récupération de l'identifiant du compte passé en paramètre
         $accountId = $request->query->get('accountId');
-        // récupération de l'identifiant de l'utilisateur courant dans le token
-        $currentUser = $this->tokenStorage->getToken()->getUser()->getId();
-        // récupération du profil passé en paramètre
-        $account = $accountRepository->find($accountId);
 
-        // si le profil n'existe pas
-        if(!$account){
-            return new JsonResponse(['message' => 'Le profil demandé n\existe pas'], JsonResponse::HTTP_NOT_FOUND);
+        // recherche du compte dans la base de données
+        $account = $this->accountRepository->find($accountId);
+
+        // contrôle si le compte existe
+        if(!$account) {
+            return new JsonResponse(['message' => 'Le compte n\'existe pas'], JsonResponse::HTTP_NOT_FOUND);
         }
 
-        // on vérifie que l'identifiant utilisateur du compte correspond bien à l'identifiant de l'utilisateur courant dans le token
-        if($account->getUser()->getId() != $currentUser) {
+        // on vérifie que l'identifiant utilisateur du compte correspond bien à l'identifiant de l'utilisateur courant dans la session
+        if($account->getUser()->getId() != $user->getId()) {
             return new JsonResponse(['message' => 'Vous n\'avez pas accès à ce compte'], JsonResponse::HTTP_FORBIDDEN);
         }
 
@@ -183,14 +231,42 @@ class ThemeController extends AbstractController
         return new JsonResponse(['themes' => $themes], JsonResponse::HTTP_OK);
     }
 
+
     #[Route('/admin/delete/{id}', name:'delete', methods:['DELETE'])]
     public function delete($id): JsonResponse
     {
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        //récupération du role de l'utilisateur
+        $currentUserRole = $session->get('user_role')[0];
+
+        // si l'utilisateur n'a pas le role admin renvoi une erreur
+        if ($currentUserRole != 'ROLE_ADMIN') {
+            return new JsonResponse(["message" => "Vous n'êtes pas autorisé à ajouter des thèmes"], JsonResponse::HTTP_FORBIDDEN);
+        }
+
+        // recherche le theme en base de données
         $theme = $this->entityManager->getRepository(Theme::class)->find($id);
+
+        //renvoi une erreur si le theme n'est pas trouvé
         if (!$theme) {
             return new JsonResponse(['message' => 'Le thème demandé n\'a pas été trouvé'], JsonResponse::HTTP_NOT_FOUND);
         }
+        // supprime le theme de la base de données
         $this->entityManager->remove($theme);
+        $this->entityManager->flush();
 
         return new JsonResponse(null, JsonResponse::HTTP_OK);
     }
@@ -199,8 +275,37 @@ class ThemeController extends AbstractController
     public function buy(
         $id,
         Request $request,
-        AccountRepository $accountRepository
     ): JsonResponse {
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        // récupération des datas envoyé par le front
+        $data = json_decode($request->getContent(), true);
+
+        // on récupére le compte à partir de la requête du front
+        $account = $this->accountRepository->find($data['profileId']);
+
+        // on vérifie que le compte existe en base de données
+        if(!$account) {
+            return new JsonResponse(['message' => 'Le compte n\'éxiste pas'], JsonResponse::HTTP_NOT_FOUND);
+        }
+
+        // on vérifie que l'identifiant utilisateur du compte correspond bien à l'identifiant de l'utilisateur courant dans le token
+        if($account->getUser()->getId() != $user->getId()) {
+            return new JsonResponse(['message' => 'Vous n\'avez pas accès à ce compte'], JsonResponse::HTTP_FORBIDDEN);
+        }
+
         // récupération du thème
         $theme = $this->themeRepository->find($id);
 
@@ -209,34 +314,15 @@ class ThemeController extends AbstractController
             return new JsonResponse(['message' => 'Le thème n\'existe pas'], JsonResponse::HTTP_NOT_FOUND);
         }
 
-        // récupération de l'identifiant de l'utilisateur courant présent dans le token
-        $currentUser = $this->tokenStorage->getToken()->getUser()->getId();
-
-        // récupération des datas envoyé par le front
-        $data = json_decode($request->getContent(), true);
-
-        // on récupére le compte à partir de la requête du front
-        $account = $accountRepository->find($data['profileId']);
-
-        // on vérifie que le compte existe en base de données
-        if(!$account) {
-            return new JsonResponse(['message' => 'Le compte n\'éxiste pas'], JsonResponse::HTTP_NOT_FOUND);
-        }
-
-        // on vérifie que l'identifiant utilisateur du compte correspond bien à l'identifiant de l'utilisateur courant dans le token
-        if($account->getUser()->getId() != $currentUser) {
-            return new JsonResponse(['message' => 'Vous n\'avez pas accès à ce compte'], JsonResponse::HTTP_FORBIDDEN);
+        // Vérifier si le compte dispose déjà du thème
+        $themes = $account->getThemeId();
+        if ($themes->contains($theme)) {
+            return new JsonResponse(['message' => 'Le compte dispose déjà de ce thème'], JsonResponse::HTTP_BAD_REQUEST);
         }
 
         // si le compte ne dispose pas de suffisament d'argent l'achat ne peut pas être effectué
         if($account->getWallet() < $theme->getValue()) {
             return new JsonResponse(['message' => 'Vous n\'avez pas assez d\'argent pour acheter ce thème'], JsonResponse::HTTP_BAD_REQUEST);
-        }
-
-        // Vérifier si le compte dispose déjà du thème
-        $themes = $account->getThemeId();
-        if ($themes->contains($theme)) {
-            return new JsonResponse(['message' => 'Le compte dispose déjà de ce thème'], JsonResponse::HTTP_BAD_REQUEST);
         }
 
         // calcul du reste du solde pour le profil
@@ -259,7 +345,7 @@ class ThemeController extends AbstractController
     {
         $datas = $this->themeRepository->findAll();
         $themes = [];
-        foreach ($datas as $data){
+        foreach ($datas as $data) {
             $theme = [
                 "id" => $data->getId(),
                 "name" => $data->getName(),
@@ -272,19 +358,37 @@ class ThemeController extends AbstractController
     #[Route("/admin/get/{id}", name:"getAdmin", methods:['GET'])]
     public function adminGetId(
         $id,
-    ): JsonResponse
-    {
-        $currentUser = $this->tokenStorage->getToken()->getUser()->getRoles()[0];
-        if ($currentUser != 'ROLE_ADMIN') {
-            return new JsonResponse(["message" => 'Vous n\êtes pas autorisé à accéder à ce contenu'],JsonResponse::HTTP_FORBIDDEN);
+    ): JsonResponse {
+        // récupération de la session
+        $session = $this->requestStack->getSession();
+
+        //récupération de l'identifiant de l'utilisateur via le cookie de session
+        $currentUserId = $session->get('user_id');
+
+        //récupération de l'utilisateur
+        $user = $this->userRepository->find($currentUserId);
+
+        //si l'utilisateur n'est pas trouvé retourner une erreur
+        if(!$user) {
+            return new JsonResponse(["message" => "L'utilisateur n'existe pas"], JsonResponse::HTTP_NOT_FOUND);
         }
 
+        //récupération du role de l'utilisateur
+        $currentUserRole = $session->get('user_role')[0];
+
+        // si l'utilisateur n'a pas le role admin renvoi une erreur
+        if ($currentUserRole != 'ROLE_ADMIN') {
+            return new JsonResponse(["message" => "Vous n'êtes pas autorisé à ajouter des thèmes"], JsonResponse::HTTP_FORBIDDEN);
+        }
+
+        // récupération du theme par sont id
         $theme = $this->themeRepository->find($id);
 
+        // si le theme n'existe pas renvoie une erreur
         if (!$theme) {
             return new JsonResponse(['message' => 'Le thème demandé n\'existe pas'], JsonResponse::HTTP_NOT_FOUND);
         }
-    
+
         $questions = [];
         foreach ($theme->getQuestionId() as $question) {
             $questionData = [
@@ -292,7 +396,7 @@ class ThemeController extends AbstractController
                 'question' => $question->getQuestion(),
                 'answers' => []
             ];
-    
+
             foreach ($question->getAnswers() as $answer) {
                 $answerData = [
                     'id' => $answer->getId(),
@@ -301,10 +405,10 @@ class ThemeController extends AbstractController
                 ];
                 $questionData['answers'][] = $answerData;
             }
-    
+
             $questions[] = $questionData;
         }
-    
+
         return new JsonResponse($questions, JsonResponse::HTTP_OK);
     }
 }
